requirements.txt
streamlit
torch
sentencepiece
pandas
scikit-learn
matplotlib
seaborn
indic-nlp-library
chardet
supabase